refactor: get soil match and ecological site similarly between TemporaryLocationCallout and LocationPrediction on dashboard
